icmbio2/requirements.txt
albumentations==1.3.0
asttokens==2.4.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
daal==2021.4.0
daal4py==2021.4.0
debugpy==1.8.0
decorator==5.1.1
efficientnet-pytorch==0.7.1
executing==2.0.1
filelock==3.9.0
fonttools==4.46.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.19.3
ipykernel==6.19.2
ipython==8.10.0
ipywidgets==8.0.4
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
mlxtend==0.21.0
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==4.1.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.2.1
scikit-learn-intelex==2021.4.0
scipy==1.10.0
seaborn==0.12.2
segmentation-models-pytorch==0.3.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchinfo==1.8.0
torchmetrics==1.2.1
torchvision==0.15.2+cu117
tornado==6.4
tqdm==4.64.1
traitlets==5.14.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.12
widgetsnbextension==4.0.9
